Ensure that Group2 array is 0-indexed and contiguous
`array_diff()` creates a sparse array. This means that it may not start
with index 0 if element 0 of the origin array is present in any other
array that it is diff'ed against. Running `array_values()` on the
resulting array will ensure that the result is a contiguous array
instead of a sparse array.