Handle mixed version groups
From time to time non-standard settings may be in use where a subset of
a group is using a different version of MediaWiki. When this happens we
will show all of the versions in use by the wikis in the group. We are
not going to try to show which particular wikis are out of step since
this is such an uncommon case, but having some idea that things are
a bit weird is helpful.

Bug: T192577